requirements.txt
apache-airflow-providers-amazon==9.16.0
apache-airflow-task-sdk==1.1.1
pandas==2.2.3
requests==2.32.5